Rework theme to utilise `sphinx-theme-builder`
Restructure the repository to be compatible with `sphinx-theme-builder`
and switch away from setuptools. This enables excluding the generated
assets from version control, since `sphinx-theme-builder` will generate
them, when generating the wheel.

Replace the usage of `yarn` with vanilla `npm`, to be more compatible
with `sphinx-theme-builder`, and for reducing the complexity in the
setup.

Explicitly declare that this theme uses the latest 14.x LTS release of
NodeJS.

Simplify `noxfile.py` since it is no longer necessary to use conda for
fetching the theme's non-Python dependencies. It happens transparently
in a `pip` based workflow. Use `stb serve` for serving documentation,
in `noxfile.py`.

Move `flake8` configuration into a flake8-specific file.