use only shield.io for badges
The services are providing different tones of green, yellow etc (github build is specifically a different green) and it hurts my OTCs. 
I simply replaced all the badges with their corresponding pure shield.io version and added a white icon for all of them (because everyone loves icons). 

Added a link to the license, that's always useful in a pypi presentation page.