DRAFT: Refactor navbar generating function.
On top of  #1771, should fix #1773 (once fininished)

This try to split the function into two parts: one that wrangle the
data,  one that generate html.

The generating the data part should be cacheable, while the other can be
made more flexible to be reusable in different part of the theme
(typically generating or not dropdowns in the sidebar or navbar).