Animate remotely loaded banners together (#1808)
The banner animation felt a little funky to me, even after the improvements made in https://github.com/pydata/pydata-sphinx-theme/pull/1693.

My hunch is that because the two banners are stacked on top of each other and the height of one affects the layout/position of the other, trying to animate the height of both of them at the same time causes the browser to stutter. Or maybe it was just because they could each load at different but often only slightly offset times. Whatever the case, I decided to do a little code clean up and change it so that they both come in together.

In the process of working on this PR it also made sense to address a TODO, and add "Version warning" to the list of translatable strings.

* pybabel extract . -F babel.cfg -o src/pydata_sphinx_theme/locale/sphinx.pot -k '_ __ l_ lazy_gettext'

* pybabel update -i src/pydata_sphinx_theme/locale/sphinx.pot -d src/pydata_sphinx_theme/locale -D sphinx

* Update src/pydata_sphinx_theme/theme/pydata_sphinx_theme/sections/announcement.html


* incorporate #1755

---------

Co-authored-by: M Bussonnier <bussonniermatthias@gmail.com>