Use navbar-nav class on text links only (#1846)
This pull request fixes #1839 while also trying to simplify the way we
use the Boostrap "navbar-nav" class.

Prior to this PR, we have both:

1. "navbar-nav" **class** applied to **text** links rendered by
"navbar-nav" **template**
2. "navbar-nav" class applied to **icon** links rendered by
"navbar-icon-links" template

This is what led to the bug in #1839 because I must have thought that
the .navbar-nav selector was for the text links, **not** the icon links.

I see two possible ways of resolving this ambiguity:

1. Keep the navbar-nav class on the text links, but remove it from the
icon links. This requires adding some style rules to the icon links so
they look the same with and without the class. This is the option shown
in this PR.
3. Keep the navbar-nav class on both sets of links, but change the name
of the navbar-nav template to something else. This would require
updating the docs, and I think it would also require theme adopters to
update their theme config files, which is why I didn't choose this
option.

---------

Co-authored-by: Daniel McCloy <dan@mccloy.info>