MAINT: Upload nightly wheels and run packaging on PRs and pushes (#1867)
It seems like it should be okay to run packaging steps on each PR and
push to `main`, but we'll see if it adds a lot of overhead. Then also
add a nightly wheel upload to
https://anaconda.org/scientific-python-nightly-wheels.

This will require getting an Anaconda token from the appropriate folks
and adding it as a secret, so marking as a draft for now.