Maint: do not use document.write
See #1920, this removes every usage of document.write in favor of
display:none with a noscript tag.

I did have to be  a little more specific for buttons as the css rule in
boostrap were overwriting the *.jsonly.

Note that this does not solves #1920 as some things (like the more
dropdown in nav bar) still require JS to work.