Additional version banner details. (#1973)
Adds a more detailed descriptions on what causes different messages to
appear in the version announcement.

This tripped me up for a little bit and I had to go searching in the
source code to find out what exactly triggered the development warning
banner to display. So this PR is in the hope of avoiding a small
headache for others!