Stop moving the Read the Docs version switcher into the sidebar (#2034)
Read the Docs recently rolled out its new "addons" project. This broke
assumptions made by code in our theme.

In particular, it means that all of the code that tries to move the Read
the Docs version switcher into the sidebar no longer works, so this pull
request removes all of that code and updates the docs.

I also removed the check for Read the Docs in the Ethical Ads template
because as far as I understand, that template is opt-in anyway, meaning
that sites that use our theme do not have that template included by
default.

Fixes https://github.com/pydata/pydata-sphinx-theme/issues/1794 and
possibly https://github.com/pydata/pydata-sphinx-theme/issues/1238.

This pull request does not address
https://github.com/pydata/pydata-sphinx-theme/issues/1933, which is
about providing a better integration with Read the Docs. The only goal
of this pull request is to remove code that is broken as a result of the
addons rollout.