DOCS - Update accessibility docs (#2039)
This pull request brings the accessibility docs up to date. 

Some of the changes in this pull request involve hard-wrapping the text
in the docs files. I feel like this itself is an accessibility
improvement as long lines are harder to read. It's also makes text
diffing in git better.

---------

Co-authored-by: Daniel McCloy <dan@mccloy.info>
Co-authored-by: Tania Allard <taniar.allard@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>