BUG - Add `--pst-color-heading` fallback (#2082)
Nick pointed in
https://github.com/pydata/pydata-sphinx-theme/pull/2058#issuecomment-2548597450
that the `--pst-heading-color` variable is used by folks to customise
the headings colour (this was the colour variable before
https://github.com/pydata/pydata-sphinx-theme/pull/2058 which brought
this in line with our naming convention `--pst-color-heading`).

This change effectively renders the use of `--pst-heading-color`
useless, so this PR adds a fallback mechanism for this variable.

I also added a note about `--pst-color-heading` being our preferred
variable.
This is not a "big" breaking change, but it will likely affect folks who
do not have their PST version pinned since I just made a release. So, it
might be worth publishing this change in a follow-up.