Add support for versionremoved directive
Note that I added a snippet in the Kitchen Sink to preview the change. I have an open PR in https://github.com/sphinx-themes/sphinx-themes.org/pull/163 to add it upstream. I can remove this snippet, if you prefer, or you can clobber it the next time you run a sync.

I'll also attach a screenshot.

- Closes #2086