Do not add HTML mixin if not building html (#2076)
When attempting to build the `latexpdf` target, pydata-sphinx-theme 0.16
currently fails on Sphinx 8.1.3 with:

```
Exception occurred:
  File "/home/simonspa/venv/lib64/python3.13/site-packages/pydata_sphinx_theme/translator.py", line 41, in visit_table
    self._table_row_indices.append(0)
    ^^^^^^^^^^^^^^^^^^^^^^^
AttributeError: 'BootstrapHTML5Translator' object has no attribute '_table_row_indices'
The full traceback has been saved in /tmp/sphinx-err-7w3z72wv.log, if you want to report the issue to the developers.
Please also report this if it was a user error, so that a better error message can be provided next time.
A bug report can be filed in the tracker at <https://github.com/sphinx-doc/sphinx/issues>. Thanks!
make: *** [Makefile:25: latexpdf] Error 1
```

The issue appears to be that the `BootstrapHTML5Mixin` is added despite
the target not being `html`. This PR moves the check for
`app.builder.format` to the top level of `setup_translators`.

(cc) @stephanlachnit