MAINT - Defer FontAwesome script to improve performance (#2209)
This pull request:

- Adds "defer" attribute to the script tag that loads the FontAwesome
library
- Consolidates our two custom icon files into one and also loads them
with defer
- Updates documentation to reflect the above two changes
- Excludes images.html from the list of Kitchen Sink pages to check with
Lighthouse.

### Defer

> [!WARNING]
> When this pull request gets released in the next version of the PyData
Sphinx Theme, it will require theme users to update their custom icon
entries in `html_js_files` to include the `defer` attribute (see the
update to the docs page in this pull request).

Without the defer attribute, the entire page has to wait for the
FontAwesome script to be downloaded and executed. Lighthouse identified
this script as the worst offender. With the defer attribute, the page
should be able to load faster.

### Excluding images.html

The reason I'm excluding images.html from Lighthouse is because it
contains images that are hotlinked from Picsum. I think that this is the
reason the Lighthouse performance score comes in too low (below 80).
Because when I examined the Lighthouse report for that page, I saw the
main reason the page failed was because of a slow Largest Contentful
Paint, and when digging into that, the report showed the following
table:

Phase | % of LCP | Timing
-- | -- | --
TTFB | 8% | 460 ms
Load Delay | 89% | 5,350 ms
Load Time | 2% | 110 ms
Render Delay | 2% | 100 ms

I assume that most of that load delay has to do with hotlinking from
Picsum and nothing that we can fix without modifying the Kitchen Sink
pages, which we're not supposed to do since those are generated from a
script.

Using the [Lighthouse score
calculator](https://googlechrome.github.io/lighthouse/scorecalc/#FCP=1665&LCP=6016&TBT=54&CLS=0&SI=2861&TTI=6045&device=mobile&version=12.1.0),
if I substract 5,000 ms from Largest Contentful Paint, the performance
score climbs to 99, which puts it in line with the new performance
scores for the other Kitchen Sink pages.

On the flip side, even though this PR excludes images.html from
Lighthouse, it actually adds more Kitchen Sink pages to the Lighthouse
audit because previously only 5 pages were being checked because the
[default for
maxAutodiscoverUrls](https://github.com/GoogleChrome/lighthouse-ci/blob/main/docs/configuration.md)
is 5.

### How I checked this PR

- [x] Looked at the results on the lighthouse-audit CI job
- Inspected the two custom icons in the preview build 
  - [x] Brave browser, desktop mode
  - [x] Brave browser, mobile mode
  - [x] Safari, desktop mode
  - [x] Firefox, desktop mode
- [x] Checked the preview build to make sure there were no new
JavaScript or other errors in the browser dev tools console log.
- [x] Checked the updated [Header Links docs
page](https://pydata-sphinx-theme--2209.org.readthedocs.build/en/2209/user_guide/header-links.html)
in the preview build

---------

Co-authored-by: Tania Allard <taniar.allard@gmail.com>